Add signed_zero to TargetFormat
